compile errors fixed
