+++ src/components/toast/test/toast.spec.ts
@@ -1,4 +1,7 @@
-import { ToastController, App, Platform, Config } from '../../../../src';
+import { App } from '../../../../src/components/app/app';
+import { Config } from '../../../../src/config/config';
+import { Platform } from '../../../../src/platform/platform';
+import { ToastController } from '../../../../src/components/toast/toast';
 
 export function run() {
