+++ src/components/virtual-scroll/test/virtual-scroll.spec.ts
@@ -2,8 +2,6 @@ import { VirtualScroll } from '../../../../src/components/virtual-scroll/virtual
 import { VirtualCell, VirtualData, VirtualNode } from '../../../../src/components/virtual-scroll/virtual-util';
 import { processRecords, populateNodeData, initReadNodes, getVirtualHeight, adjustRendered } from '../../../../src/components/virtual-scroll/virtual-util';
 
-
-
 export function run() {
 
 describe('VirtualScroll', () => {
