--- src/platform/storage.ts
@@ -1,3 +1,0 @@
-export * from './storage/storage'
-export * from './storage/local-storage'
-export * from './storage/sql'
