+++ src/util/mock-providers.ts
@@ -1,16 +1,28 @@
 import { ChangeDetectorRef, ElementRef, NgZone, Renderer } from '@angular/core';
 import { Location } from '@angular/common';
 
-import { App, Config, DeepLinker, Form, GestureController, Keyboard, NavLinkConfig, Platform, Tab, Tabs, Transition, UrlSerializer, ViewController } from '../../src';
-import { NavControllerBase } from '../../src/components/nav/nav-controller-base';
+import { App } from '../components/app/app';
+import { Config } from '../config/config';
+import { DeepLinker } from '../navigation/deep-linker';
+import { Form } from './form';
+import { GestureController } from '../gestures/gesture-controller';
+import { Keyboard } from './keyboard';
+import { Platform } from '../platform/platform';
+import { Tab }  from '../components/tabs/tab';
+import { Tabs }  from '../components/tabs/tabs';
+import { Transition } from '../transitions/transition';
+import { UrlSerializer } from '../navigation/url-serializer';
+import { ViewController } from '../navigation/view-controller';
+
+import { NavControllerBase } from '../navigation/nav-controller-base';
 
 
 export const mockConfig = function(config?: any) {
-  return new Config(config);
+  return new Config();
 };
 
 export const mockPlatform = function(platforms?: string[]) {
-  return new Platform(platforms);
+  return new Platform();
 };
 
 export const mockApp = function(config?: Config, platform?: Platform) {
@@ -81,7 +93,7 @@ export const mockNavController = function(): NavControllerBase {
   let platform = mockPlatform();
 
   let config = mockConfig();
-  config.setPlatform(platform);
+  //config.setPlatform(platform);
 
   let app = mockApp(config, platform);
 
@@ -99,9 +111,9 @@ export const mockNavController = function(): NavControllerBase {
 
   let gestureCtrl = new GestureController(app);
 
-  let navLikConfig = new NavLinkConfig([]);
+  //let navLikConfig = new NavLinkConfig([]);
 
-  let serializer = new UrlSerializer(navLikConfig);
+  let serializer = new UrlSerializer(null);
 
   let location = mockLocation();
 
@@ -117,6 +129,7 @@ export const mockNavController = function(): NavControllerBase {
     renderer,
     compiler,
     gestureCtrl,
+    null,
     deepLinker
   );
 };
@@ -125,7 +138,7 @@ export const mockTab = function(parentTabs: Tabs): Tab {
   let platform = mockPlatform();
 
   let config = mockConfig();
-  config.setPlatform(platform);
+  //config.setPlatform(platform);
 
   let app = (<any>parentTabs)._app || mockApp(config, platform);
 
@@ -145,9 +158,9 @@ export const mockTab = function(parentTabs: Tabs): Tab {
 
   let gestureCtrl = new GestureController(app);
 
-  let navLikConfig = new NavLinkConfig([]);
+  //let navLikConfig = new NavLinkConfig([]);
 
-  let serializer = new UrlSerializer(navLikConfig);
+  let serializer = new UrlSerializer(null);
 
   let location = mockLocation();
 
@@ -164,6 +177,7 @@ export const mockTab = function(parentTabs: Tabs): Tab {
     compiler,
     changeDetectorRef,
     gestureCtrl,
+    null,
     linker
   );
